\name{apTreeshape-internal}
\alias{as.treeshape.treebalance}
\alias{is.binary.phylo}
\alias{treebalance}


\title{Internal apTreeshape Functions}

\description{
  Internal apTreeshape functions.
}

\note{
  These are not to be called by the user (or in some cases are just waiting for proper documentation to be written).
}

\keyword{internal}
