\name{depth}
\alias{depth}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal BetaAlphaEta function
}
\description{
Gives the node depths of the tree, in the order of node labels 
}
\usage{
depth(tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
A phylo object
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}