\name{insert}
\alias{insert}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Insert an element in a vector
}
\description{
Insert an element in a vector
}
\usage{
insert(v, e, i, replace)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{v}{
Vector}
  \item{e}{
Element to be inserted
}
  \item{i}{
Position where the element should be inserted
}
  \item{replace}{
Should the elment at position i be replaced by e?
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}
