\name{nodes_depths_ordonnes}
\alias{nodes_depths_ordonnes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Gets the node depths of the tree
}
\description{
Gives the node depths of the tree, in the order needed to reconstruct the tree using build.tree
 }
\usage{
nodes_depths_ordonnes(tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
A phylo object
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}