\name{simulate.R.K}
\alias{simulate.R.K}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal BetaAlphaEta function
}
\description{
Simulate interval split together with the number of tips in each resulting subinterval (see appendix 1)
}
\usage{
simulate.R.K(beta, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{
Imbalance index
}
  \item{n}{
Tip number in the parent clade
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}