\name{simulate.R}
\alias{simulate.R}
\title{
Internal BetaAlphaEta function
}
\description{
Simulate interval split conditioned to the number of tips in each resulting subinterval (see appendix 1)
}
\usage{
simulate.R(beta, n, K)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{
Imbalance index
}
  \item{n}{
Tip number in the parent clade
}
  \item{K}{
Tip number in the left daugther clade
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}