\name{simulate.Yi}
\alias{simulate.Yi}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal BetaAlphaEta function
}
\description{
Simulates the random variables Y_i (see online appendix)
}
\usage{
simulate.Yi(N, epsilon, beta, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
Number of simulated Y_i
}
  \item{epsilon}{
Minimum size of unsampled splits (see appendix 1)
}
  \item{beta}{
Imbalance index
}
  \item{n}{
Number of marks in the interval
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}

