\name{split}
\alias{split}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Computes the splits at each node of a tree
}
\description{
Computes the splits at each node of a tree
}
\usage{
split(tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
A phylo object
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}