\name{yule}
\alias{yule}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal BetaAlphaEta function
}
\description{
Simulates a phylogeny in a birth-death model, conditionned by the number of tips and the maximum age of the root
}
\usage{
yule(N, b, d, tmax)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
Tip number
}
  \item{b}{
Birth rate
}
  \item{d}{
Death rate
}
  \item{tmax}{
Maximum root depth
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}