\name{a_N}
\alias{a_N}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Computes a_n
}
\description{
Values of a_n for n in 1:N (see appendix 1)
}
\usage{
a_N(beta, N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{
Imbalance index
}
  \item{N}{
Tip number
}
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}
