\name{apTreeshape-internal}
\alias{as.treeshape.treebalance}
\alias{is.binary.phylo}
\alias{pandit}
\alias{raldous}
\alias{rbiased}
\alias{rpda}
\alias{ryule}
\alias{treebalance}
\alias{treebase}
\alias{read.tree2}
\alias{logratio}
\alias{Delta}
\alias{ancestor}
\alias{smaller.clade.spectrum2}



\title{Internal apTreeshape Functions}

\description{
  Internal apTreeshape functions.
}

\note{
  These are not to be called by the user (or in some cases are just waiting for proper documentation to be written).
}

\keyword{internal}
