\name{simulate.R.K}
\alias{simulate.R.K}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulate (R,K)
}
\description{
Simulate a split and the number of marks in each resulting subintervals
}
\usage{
simulate.R.K(beta, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{
Imbalance index
}
  \item{n}{
Tip number in the parent clade
}
}
\references{
Maliet O., Gascuel F., Lambert A. (2018) \emph{Ranked tree shapes, non-random 
extinctions and the loss of phylogenetic diversity}, bioRxiv 224295, 
doi: https://doi.org/10.1101/224295 
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}