\name{simulate.R}
\alias{simulate.R}
\title{
Simulate R knowing K=k
}
\description{
Simulate a split conditioned to the number of marks in each resulting subinterval
}
\usage{
simulate.R(beta, n, K)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{
Imbalance index
}
  \item{n}{
Tip number in the parent clade
}
  \item{K}{
Tip number in the left daugther clade
}
}
\references{
Maliet O., Gascuel F., Lambert A. (2018) \emph{Ranked tree shapes, non-random 
extinctions and the loss of phylogenetic diversity}, bioRxiv 224295, 
doi: https://doi.org/10.1101/224295 
}
\author{
Odile Maliet, Fanny Gascuel & Amaury Lambert
}